refactor: algolia language dropdown
