+++ libs/journeys/ui/src/components/SearchBar/AlgoliaLanguageDropdown/index.ts
@@ -1,0 +1,1 @@
+export { AlgoliaLanguageDropdown } from './AlgoliaLanguageDropdown'
